+++ CHANGELOG.md
@@ -33,6 +33,7 @@ Given the `MAJOR.MINOR.PATCH` pattern, here is how we decide to increment:
 - Added `Object.defineProperty` polyfill.
 - Added unit test for `aria-state.js`.
 - Wagtail CMS
+- Added `gulp test:a11y` accessibility testing using node-wcag.
 
 ### Changed
 - Updated Video Code to make it usable on Events pages.
@@ -49,7 +50,7 @@ Given the `MAJOR.MINOR.PATCH` pattern, here is how we decide to increment:
 - Server port is now at 8000
 - Updated Dep Dir title to include "Acting"
 - included with context flag for macros that make a call to request object
-
+- Added `binaryDirectory` parameter to `fsHelper.getBinary` helper function.
 
 ### Removed
 - Disables tests for landing page events, since we don't currently have events.
@@ -64,6 +65,7 @@ Given the `MAJOR.MINOR.PATCH` pattern, here is how we decide to increment:
 - Fixed instructions for gulp watch
 - New way to run the server documented in the INSTALL.MD
 - New way to define url routing, no longer automatically set by file path
+
 
 ## 3.0.0-2.3.0 - 2015-08-27